Code formatting
One click IDE action

+++ src/main/java/minefantasy/mf2/api/armour/ItemArmourMFBase.java
@@ -18,9 +18,9 @@ import net.minecraft.util.StatCollector;
 import net.minecraftforge.common.ISpecialArmor;
 import net.minecraftforge.common.util.EnumHelper;
 
-public class ItemArmourMFBase extends ItemArmor implements ISpecialArmor, IArmourMF, IArmourRating, ISpecialArmourMF
-{
-	public static ArmorMaterial baseMaterial = EnumHelper.addArmorMaterial("MF Armour Base", 0, new int[]{2, 6, 5, 2}, 0);
+public class ItemArmourMFBase extends ItemArmor implements ISpecialArmor, IArmourMF, IArmourRating, ISpecialArmourMF {
+	public static ArmorMaterial baseMaterial = EnumHelper.addArmorMaterial("MF Armour Base", 0,
+			new int[] { 2, 6, 5, 2 }, 0);
 	private int piece;
 	private int baseRating;
 	public float baseAR;
@@ -31,11 +31,10 @@ public class ItemArmourMFBase extends ItemArmor implements ISpecialArmor, IArmou
 	public ArmourDesign design;
 	protected float suitBulk;
 	public float DT;
-	
+
 	public static final DecimalFormat decimal_format = new DecimalFormat("#.#");
-	
-	public ItemArmourMFBase(String name, ArmourMaterialMF material, ArmourDesign AD, int slot, String tex)
-	{
+
+	public ItemArmourMFBase(String name, ArmourMaterialMF material, ArmourDesign AD, int slot, String tex) {
 		super(baseMaterial, 0, slot);
 		this.material = material;
 		baseAR = material.baseAR;
@@ -45,256 +44,231 @@ public class ItemArmourMFBase extends ItemArmor implements ISpecialArmor, IArmou
 		design = AD;
 		suitBulk = design.getBulk();
 		texture = tex;
-		float baseDura = material.durability * design.getDurability()/2F;
-		float dura = baseDura /2F + (baseDura /2F * ArmourCalculator.sizes[slot] / ArmourCalculator.sizes[1]);
-		this.setMaxDamage((int)dura);
+		float baseDura = material.durability * design.getDurability() / 2F;
+		float dura = baseDura / 2F + (baseDura / 2F * ArmourCalculator.sizes[slot] / ArmourCalculator.sizes[1]);
+		this.setMaxDamage((int) dura);
 		this.setUnlocalizedName(name);
-		setDT(getBaseAC()*scalePiece());
-		
-		baseRating = ArmourCalculator.translateToVanillaAR(getBaseAC()+1F, baseMaterial.getDamageReductionAmount(slot), 15);
+		setDT(getBaseAC() * scalePiece());
+
+		baseRating = ArmourCalculator.translateToVanillaAR(getBaseAC() + 1F,
+				baseMaterial.getDamageReductionAmount(slot), 15);
 	}
-	public ItemArmourMFBase setDT(float DT)
-	{
+
+	public ItemArmourMFBase setDT(float DT) {
 		this.DT = DT;
 		return this;
 	}
-	/*  Piece | Slot
-	 *  0     | 3
-	 *  1     | 2
-	 *  2     | 1
-	 *  3     | 0 */
+
+	/*
+	 * Piece | Slot 0 | 3 1 | 2 2 | 1 3 | 0
+	 */
 	@Override
-	public void damageArmor(EntityLivingBase entity, ItemStack stack, DamageSource source, int damage, int slot)
-	{
-		initArmourDamage(entity, stack, (float)damage);
+	public void damageArmor(EntityLivingBase entity, ItemStack stack, DamageSource source, int damage, int slot) {
+		initArmourDamage(entity, stack, damage);
 	}
-	public void initArmourDamage(EntityLivingBase user, ItemStack armour, float damage)
-	{
-		armour.damageItem((int)(damage/5F)+1, user);
+
+	public void initArmourDamage(EntityLivingBase user, ItemStack armour, float damage) {
+		armour.damageItem((int) (damage / 5F) + 1, user);
 	}
 
 	@Override
-	public int getArmorDisplay(EntityPlayer player, ItemStack armour, int slot)
-	{
-		float max = (float)armour.getMaxDamage();
-		float dam = max - (float)armour.getItemDamage();
-		
-		return (int)Math.round(5F / max * dam);
+	public int getArmorDisplay(EntityPlayer player, ItemStack armour, int slot) {
+		float max = armour.getMaxDamage();
+		float dam = max - armour.getItemDamage();
+
+		return Math.round(5F / max * dam);
 	}
 
 	@Override
-	public ArmorProperties getProperties(EntityLivingBase player, ItemStack armour, DamageSource source, double damage, int slot) 
-	{
+	public ArmorProperties getProperties(EntityLivingBase player, ItemStack armour, DamageSource source, double damage,
+			int slot) {
 		float AC = getProtectionRatio(armour);
 		AC = ArmourCalculator.getArmourValueMod(armour, AC);
-		
-		if(ArmourCalculator.advancedDamageTypes && !player.worldObj.isRemote)
-		{
-			AC = ArmourCalculator.adjustACForDamage(source, AC, getProtectiveTrait(armour, 0), getProtectiveTrait(armour, 1), getProtectiveTrait(armour, 2));
+
+		if (ArmourCalculator.advancedDamageTypes && !player.worldObj.isRemote) {
+			AC = ArmourCalculator.adjustACForDamage(source, AC, getProtectiveTrait(armour, 0),
+					getProtectiveTrait(armour, 1), getProtectiveTrait(armour, 2));
 		}
-		
+
 		AC *= getACModifier(player, armour, source, damage);
-		
-		if(source == DamageSource.fall && design == ArmourDesign.PLATE)
-		{
-			AC = 1F/material.armourWeight;
-		}
-		else if(source.isMagicDamage())
-		{
+
+		if (source == DamageSource.fall && design == ArmourDesign.PLATE) {
+			AC = 1F / material.armourWeight;
+		} else if (source.isMagicDamage()) {
 			AC = getMagicAC(AC, source, damage, player);
-		}
-		else if(source == DamageSource.onFire)
-		{
+		} else if (source == DamageSource.onFire) {
 			AC *= getACForBurn();
-		}
-		else if(source.isUnblockable())
-		{
+		} else if (source.isUnblockable()) {
 			AC *= getUnblockableResistance(armour, source);
 		}
 		AC *= getSpecialModifier(armour, source);
-		
-		if(player.getEntityData().hasKey("MF_ZombieArmour"))
-		{
+
+		if (player.getEntityData().hasKey("MF_ZombieArmour")) {
 			AC -= 1.5F;
-			if(AC < 0)AC = 0;
+			if (AC < 0)
+				AC = 0;
 		}
-		AC++;//because 1.0AC = no armour so it adds ontop of this
-		
+		AC++;// because 1.0AC = no armour so it adds ontop of this
+
 		double totalPercent = ArmourCalculator.convertToPercent(AC);
-		double maxPercent = 0.99D;//max percentage is 99% damage to avoid immunity
-		
-		if(totalPercent > maxPercent)
-		{
+		double maxPercent = 0.99D;// max percentage is 99% damage to avoid immunity
+
+		if (totalPercent > maxPercent) {
 			totalPercent = maxPercent;
 		}
-		double percent = totalPercent*scalePiece();
-		if(percent < 0)
-		{
+		double percent = totalPercent * scalePiece();
+		if (percent < 0) {
 			percent = 0;
 		}
-		
+
 		return new ArmorProperties(0, percent, Integer.MAX_VALUE);
 	}
-	
+
 	protected float getSpecialModifier(ItemStack armour, DamageSource source) {
 		return 1.0F;
 	}
-	private float getACForBurn() 
-	{
-		return armourWeight >= ArmourCalculator.encumberanceArray[1] ? 0.1F : armourWeight >= ArmourCalculator.encumberanceArray[0] ? 0.05F : 0F;
+
+	private float getACForBurn() {
+		return armourWeight >= ArmourCalculator.encumberanceArray[1] ? 0.1F
+				: armourWeight >= ArmourCalculator.encumberanceArray[0] ? 0.05F : 0F;
 	}
-	public float getUnblockableResistance(ItemStack item, DamageSource source)
-	{
+
+	public float getUnblockableResistance(ItemStack item, DamageSource source) {
 		return 0F;
 	}
-	
-	public float getMagicAC(float AC, DamageSource source, double damage, EntityLivingBase player)
-	{
+
+	public float getMagicAC(float AC, DamageSource source, double damage, EntityLivingBase player) {
 		return 0F;
 	}
-	protected float getACModifier(EntityLivingBase player, ItemStack armour, DamageSource source, double damage) 
-	{
+
+	protected float getACModifier(EntityLivingBase player, ItemStack armour, DamageSource source, double damage) {
 		return 1.0F;
 	}
-	public float scalePiece()
-	{
+
+	public float scalePiece() {
 		return ArmourCalculator.sizes[piece];
 	}
 
-	protected float getProtectionRatio(ItemStack item) 
-	{
+	protected float getProtectionRatio(ItemStack item) {
 		return getBaseAC();
 	}
-	protected float getBaseAC() 
-	{
-		return material.baseAR*design.getRating();
+
+	protected float getBaseAC() {
+		return material.baseAR * design.getRating();
 	}
 
-	private void addModifier(List list, ItemStack item, float ratio, String name)
-	{
-		if(getProtectionRatio(item) != ratio)
-        {
-        	if(getProtectionRatio(item) > ratio)
-        	{
-        		float percent = (ratio / getProtectionRatio(item))-1F;
-                list.add(EnumChatFormatting.RED+
-                        StatCollector.translateToLocalFormatted("attribute.modifier.take."+ 1,
-                                decimal_format.format(-percent * 100),
-                                        StatCollector.translateToLocal("attribute.armour."+name)));
-            }else
-            {
-            	float percent = (ratio / getProtectionRatio(item))-1F;
-                list.add(EnumChatFormatting.DARK_GREEN+
-                        StatCollector.translateToLocalFormatted("attribute.modifier.plus."+ 1,
-                                decimal_format.format(percent * 100),
-                                        StatCollector.translateToLocal("attribute.armour."+name)));
-            }
-        }
+	private void addModifier(List list, ItemStack item, float ratio, String name) {
+		if (getProtectionRatio(item) != ratio) {
+			if (getProtectionRatio(item) > ratio) {
+				float percent = (ratio / getProtectionRatio(item)) - 1F;
+				list.add(EnumChatFormatting.RED + StatCollector.translateToLocalFormatted(
+						"attribute.modifier.take." + 1, decimal_format.format(-percent * 100),
+						StatCollector.translateToLocal("attribute.armour." + name)));
+			} else {
+				float percent = (ratio / getProtectionRatio(item)) - 1F;
+				list.add(EnumChatFormatting.DARK_GREEN + StatCollector.translateToLocalFormatted(
+						"attribute.modifier.plus." + 1, decimal_format.format(percent * 100),
+						StatCollector.translateToLocal("attribute.armour." + name)));
+			}
+		}
 	}
-	public ArmourMaterialMF getMaterial()
-	{
+
+	public ArmourMaterialMF getMaterial() {
 		return this.material;
 	}
-	
+
 	@Override
-	public int getMaxDamage(ItemStack stack)
-	{
-    	if(isUnbreakable())
-		{
-    		setUnbreakable(stack);
+	public int getMaxDamage(ItemStack stack) {
+		if (isUnbreakable()) {
+			setUnbreakable(stack);
 		}
 		return ToolHelper.setDuraOnQuality(stack, super.getMaxDamage());
 	}
 
-	protected boolean isUnbreakable()
-	{
+	protected boolean isUnbreakable() {
 		return false;
 	}
 
 	@Override
-	public float scalePiece(ItemStack item) 
-	{
+	public float scalePiece(ItemStack item) {
 		return scalePiece();
 	}
-	
+
 	@Override
-	public int getItemEnchantability()
-	{
+	public int getItemEnchantability() {
 		return material.enchantment;
 	}
 
 	@Override
-	public float getPieceWeight(ItemStack item, int slot) 
-	{
+	public float getPieceWeight(ItemStack item, int slot) {
 		return armourWeight * ArmourCalculator.sizes[slot];
 	}
-	
-	public static void setUnbreakable(ItemStack tool)
-	{
-		if(!tool.hasTagCompound())
-		{
+
+	public static void setUnbreakable(ItemStack tool) {
+		if (!tool.hasTagCompound()) {
 			tool.setTagCompound(new NBTTagCompound());
 		}
 		tool.getTagCompound().setBoolean("Unbreakable", true);
 	}
-	
-	private static final float[] AC = new float[]{40F, 60F};
+
+	private static final float[] AC = new float[] { 40F, 60F };
 
 	@Override
-	public String getSuitWeigthType(ItemStack item)
-	{
+	public String getSuitWeigthType(ItemStack item) {
 		return design.getGroup();
 	}
+
 	@Override
-	public float getDTValue(EntityLivingBase user, ItemStack armour, DamageSource src)
-	{
+	public float getDTValue(EntityLivingBase user, ItemStack armour, DamageSource src) {
 		float DT2 = DT;
-		
-		if(ArmourCalculator.advancedDamageTypes && !user.worldObj.isRemote)
-		{
-			DT2 = ArmourCalculator.adjustACForDamage(src, DT2, getProtectiveTrait(armour, 0), getProtectiveTrait(armour, 1), getProtectiveTrait(armour, 2));
+
+		if (ArmourCalculator.advancedDamageTypes && !user.worldObj.isRemote) {
+			DT2 = ArmourCalculator.adjustACForDamage(src, DT2, getProtectiveTrait(armour, 0),
+					getProtectiveTrait(armour, 1), getProtectiveTrait(armour, 2));
 		}
 		return DT2;
 	}
+
 	@Override
 	@SideOnly(Side.CLIENT)
-	public float getDTDisplay(ItemStack armour, int damageType)
-	{
-		return design.getProtectiveTraits()[damageType]*DT;
+	public float getDTDisplay(ItemStack armour, int damageType) {
+		return design.getProtectiveTraits()[damageType] * DT;
 	}
+
 	@Override
-	public float getDRValue(EntityLivingBase user, ItemStack armour, DamageSource src)
-	{
-		float DR = getProtectionRatio(armour)*scalePiece();
-		
-		if(ArmourCalculator.advancedDamageTypes && !user.worldObj.isRemote)
-		{
-			DR = ArmourCalculator.adjustACForDamage(src, DR, getProtectiveTrait(armour, 0), getProtectiveTrait(armour, 1), getProtectiveTrait(armour, 2));
+	public float getDRValue(EntityLivingBase user, ItemStack armour, DamageSource src) {
+		float DR = getProtectionRatio(armour) * scalePiece();
+
+		if (ArmourCalculator.advancedDamageTypes && !user.worldObj.isRemote) {
+			DR = ArmourCalculator.adjustACForDamage(src, DR, getProtectiveTrait(armour, 0),
+					getProtectiveTrait(armour, 1), getProtectiveTrait(armour, 2));
 		}
 		return DR;
 	}
+
 	@Override
 	@SideOnly(Side.CLIENT)
-	public float getDRDisplay(ItemStack armour, int damageType)
-	{
-		float DR = getProtectionRatio(armour)*scalePiece();
-		
-		if(ArmourCalculator.advancedDamageTypes)
-		{
-			DR = ArmourCalculator.modifyACForType(damageType, DR, getProtectiveTrait(armour, 0), getProtectiveTrait(armour, 1), getProtectiveTrait(armour, 2));
+	public float getDRDisplay(ItemStack armour, int damageType) {
+		float DR = getProtectionRatio(armour) * scalePiece();
+
+		if (ArmourCalculator.advancedDamageTypes) {
+			DR = ArmourCalculator.modifyACForType(damageType, DR, getProtectiveTrait(armour, 0),
+					getProtectiveTrait(armour, 1), getProtectiveTrait(armour, 2));
 		}
 		return DR;
 	}
-	public float getProtectiveTrait(ItemStack item, int dtype) 
-	{
+
+	public float getProtectiveTrait(ItemStack item, int dtype) {
 		return design.getProtectiveTraits()[dtype];
 	}
+
 	/**
 	 * Gets a modifier, improving resistances against the elements
-	 * @param hazard is the type of damage, like "fire"
+	 * 
+	 * @param hazard
+	 *            is the type of damage, like "fire"
 	 */
-	public float getResistanceModifier(ItemStack item, String hazard)
-	{
+	public float getResistanceModifier(ItemStack item, String hazard) {
 		return 1.0F;
 	}
 }
